Be more careful with fd_set.
These changes should avoid having some weird 100%-CPU loop if a module
stumbles at the startup (see the former MPD bug).

The problem is that as soon as MPD stops, the MPD module can no longer
access it, and starts going weird (I need to figure this out).

This patch should allow to stop the CPU madness by simply restarting
dzenstat; note that this is a workaround, not a solution.